feat(complete)!: Open to new shells for dynamic completions
