+++ clap_complete/examples/dynamic.rs
@@ -16,14 +16,14 @@ fn command() -> clap::Command {
                 .value_parser(["json", "yaml", "toml"]),
         )
         .args_conflicts_with_subcommands(true);
-    clap_complete::dynamic::bash::CompleteCommand::augment_subcommands(cmd)
+    clap_complete::dynamic::shells::CompleteCommand::augment_subcommands(cmd)
 }
 
 fn main() {
     let cmd = command();
     let matches = cmd.get_matches();
     if let Ok(completions) =
-        clap_complete::dynamic::bash::CompleteCommand::from_arg_matches(&matches)
+        clap_complete::dynamic::shells::CompleteCommand::from_arg_matches(&matches)
     {
         completions.complete(&mut command());
     } else {
