+++ clap_complete/src/dynamic/mod.rs
@@ -2,7 +2,6 @@
 
 mod completer;
 
-pub mod bash;
 pub mod shells;
 
 pub use completer::*;
